+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 _(add items here for easier creation of next log entry)_
 
+- Add ability to specify a `defaultValue` to prefill the input.
+
 ## 0.3.0 - 2017-03-09
 
 - [Breaking] Add `displayMenu` property. The default is `inline` which was not the previous default.